fix(chat): move layout to components, move marketplaceHeader to Marketplace
