+++ apps/chat/src/pages/marketplace/index.tsx
@@ -1,8 +1,8 @@
 import { getCommonPageProps } from '@/src/utils/server/get-common-page-props';
 
-import MarketplaceHeader from '@/src/components/Header/MarketplaceHeader';
 import MarketplaceView from '@/src/components/Marketplace/Marketplace';
 import MarketplaceFilterbar from '@/src/components/Marketplace/MarketplaceFilterbar';
+import MarketplaceHeader from '@/src/components/Marketplace/MarketplaceHeader';
 
 export default function Marketplace() {
   return (
